1.添加T接头相关算法 移除之前立板层余高算法 2.修复焊接长度超过32768的bug

+++ App/WeldAPI/SysMath.h
@@ -28,7 +28,7 @@ public:
     //角度转换
     float angel;
     //余高
-    int reinforcementValue;
+    float reinforcementValue;
     //溶敷系数 *100
     int meltingCoefficientValue;
     //打底层限制条件
